Fix QML support (#75)
* Check if index in negative before incrementing

This one is related to https://github.com/101arrowz/fflate/issues/71 :) After the `setTiemout` issues was sorted I found antoher one.

In the hMap function there is a moment where the program may try to access a negative index. Although in Node.js and browser that does not raise an error, in QML it does raise a TypeError.

Good thing is after this fix it all works over there. 

I was not able to run the complete fflate test suite in my machine, so hopefully this works fine.

* Update src/index.ts

Co-authored-by: Manu MA <manu.valladolid@gmail.com>

* Use implicit comparison to zero

Co-authored-by: Manu MA <manu.valladolid@gmail.com>
Co-authored-by: 101arrowz <arjunbarrett@gmail.com>